# openapi/openapi.yml
openapi: 3.0.0
info:
  title: Auth API
  version: 1.0.2
  description: |
    Authentication and Authorization API for E-commerce platform.
    Supports admin panel users with RBAC.
    Uses PASETO tokens for secure authentication.
paths:
...
  /v1/admin/login:
    post:
      summary: Admin user login
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Invalid credentials
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Account disabled
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too many login attempts
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/me:
    get:
      summary: Get current admin user profile
      operationId: adminGetProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Admin user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProfile"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # ===========================================================================
  # Admin User Management (requires super_admin or user_manager role)
  # ===========================================================================
  /v1/admin/users/create:
    post:
      summary: Create a new admin user
      operationId: adminUserCreate
      security:
        - BearerAuth: [users:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreateRequest"
      responses:
        "201":
          description: Admin user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden - insufficient permissions
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Email already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/users/get/{id}:
    get:
      summary: Get admin user by ID
      operationId: adminUserGetById
      security:
        - BearerAuth: [users:read]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Admin user details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/users/list:
    get:
      summary: List admin users with pagination
      operationId: adminUserList
      security:
        - BearerAuth: [users:read]
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
            example: 20
        - name: role
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/AdminRole"
        - name: enabled
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
          description: Search by name or email
      responses:
        "200":
          description: List of admin users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "400":
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/users/disable/{id}:
    post:
      summary: Disable an admin user
      operationId: adminUserDisable
      security:
        - BearerAuth: [users:delete]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User disabled
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden - cannot disable yourself or super_admin
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/admin/users/enable/{id}:
    post:
      summary: Enable an admin user
      operationId: adminUserEnable
      security:
        - BearerAuth: [users:write]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User enabled
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: User not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # ===========================================================================
  # Token Management
  # ===========================================================================
  /v1/admin/logout:
    post:
      summary: Logout current user (invalidate refresh token)
      operationId: adminLogout
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Logged out successfully
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/token/refresh:
    post:
      summary: Refresh access token using refresh token
      operationId: tokenRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Invalid or expired refresh token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

  # ===========================================================================
  # Roles & Permissions (informational)
  # ===========================================================================
  /v1/roles:
    get:
      summary: Get all available roles
      operationId: getRoles
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolesResponse"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"

# =============================================================================
# Components
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: PASETO token in Authorization header
  schemas:
...
    AdminRole:
      type: string
      description: Admin user role (configured via application settings)
      example: "super_admin"
      pattern: "^[a-z_]+$"
      minLength: 2
      maxLength: 50
    Permission:
      type: string
      description: System permission in format "resource:action"
      example: "products:read"
      pattern: "^[a-z_]+:[a-z_]+$"
      minLength: 3
      maxLength: 50

    # =========================================================================
    # Auth Schemas
    # =========================================================================
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@example.com"
          minLength: 5
          maxLength: 255
          pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
        password:
          type: string
          format: password
          example: "SecureP@ss123"
          minLength: 1
          maxLength: 72
    AdminAuthResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - expiresAt
        - refreshExpiresIn
        - refreshExpiresAt
        - tokenType
        - user
      properties:
        accessToken:
          type: string
          description: PASETO access token
        refreshToken:
          type: string
          description: Refresh token for obtaining new access tokens
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
          example: 900
        expiresAt:
          type: string
          format: date-time
          description: Access token absolute expiration time (ISO 8601 UTC)
          example: "2026-01-28T12:15:00Z"
        refreshExpiresIn:
          type: integer
          description: Refresh token expiration time in seconds
          example: 604800
        refreshExpiresAt:
          type: string
          format: date-time
          description: Refresh token absolute expiration time (ISO 8601 UTC)
          example: "2026-02-04T12:00:00Z"
        tokenType:
          type: string
          example: "Bearer"
        user:
          $ref: "#/components/schemas/AdminUserProfile"

    # =========================================================================
    # Admin User Schemas
    # =========================================================================
    AdminUserProfile:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - permissions
        - enabled
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/AdminRole"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
          description: Effective permissions based on role
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    AdminUserCreateRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - role
      properties:
        email:
          type: string
          format: email
          minLength: 5
          maxLength: 255
          pattern: "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).{8,}$"
          description: Must contain at least 8 characters, including uppercase, lowercase, and a number
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[\\p{L}\\s'-]+$"
          description: Letters, spaces, hyphens and apostrophes only
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          pattern: "^[\\p{L}\\s'-]+$"
          description: Letters, spaces, hyphens and apostrophes only
        role:
          $ref: "#/components/schemas/AdminRole"
    AdminUserResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - enabled
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          $ref: "#/components/schemas/AdminRole"
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    AdminUserListResponse:
      type: object
      required:
        - items
        - page
        - size
        - total
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserResponse"
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total:
          type: integer
          example: 100

    # =========================================================================
    # Token Schemas
    # =========================================================================
    TokenRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: The refresh token
...
    RoleInfo:
      type: object
      required:
        - name
        - description
        - permissions
      properties:
        name:
          $ref: "#/components/schemas/AdminRole"
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    RolesResponse:
      type: object
      required:
        - roles
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleInfo"

    # =========================================================================
    # Error Schemas
    # =========================================================================
    Problem:
      type: object
      description: RFC7807 Problem Details
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: "https://api.example.com/errors/validation"
        title:
          type: string
          example: "Validation Error"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The request body contains invalid data"
        instance:
          type: string
          format: uri
        traceId:
          type: string
          example: "abc123"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Invalid email format"